+++ optimacore/gui.py
@@ -6,7 +6,7 @@ Note: Callback functions cannot be easily decorated, so logging is applied per m
 """
 
 from optimacore.system import logUsage, accepts, returns, logger, SystemSettings
-from optimacore.framework_settings import FrameworkSettings, DatabookSettings
+from optimacore.structure_settings import FrameworkSettings, DatabookSettings
 from optimacore.framework_io import FrameworkTemplateInstructions, createFrameworkTemplate
 from optimacore.framework import ProjectFramework
 #from optimacore.databook_settings import DatabookSettings