+++ tests/test.py
@@ -7,11 +7,10 @@ Tests the creation, importing and exporting of model framework files.
 import unittest
 
 from optimacore.project import Project
-from optimacore.framework_io import createFrameworkTemplate
 
 from optimacore.workbook import writeWorkbook, readWorkbook
 from optimacore.system import SystemSettings as SS
-from optimacore.framework_settings import FrameworkSettings as FS
+from optimacore.structure_settings import FrameworkSettings as FS
 
 import sys
 import pprint
@@ -34,16 +33,12 @@ class EverythingTest(unittest.TestCase):
     #    self.proj.createDatabook(databook_path = self.databook_empty_filepath)
     #    return None
 
-    def test_template_process(self):
+    def test_workbook_process(self):
         """  """
         writeWorkbook(workbook_path = "./frameworks/framework_test.xlsx", framework = None, data = None, instructions = None, workbook_type = SS.WORKBOOK_KEY_FRAMEWORK)
         readWorkbook(workbook_path = "./frameworks/framework_test.xlsx", framework = self.proj.framework, data = None, workbook_type = SS.WORKBOOK_KEY_FRAMEWORK)
         writeWorkbook(workbook_path = "./databooks/databook_test.xlsx", framework = self.proj.framework, data = None, instructions = None, workbook_type = SS.WORKBOOK_KEY_DATA)
         readWorkbook(workbook_path = "./databooks/databook_test.xlsx", framework = self.proj.framework, data = self.proj.data, workbook_type = SS.WORKBOOK_KEY_DATA)
-        #createFrameworkTemplate(framework_path = self.framework_template_filepath)
-        #self.proj.getFramework().importFromFile(framework_path = self.framework_template_filepath)
-        #self.proj.createDatabook(databook_path = self.databook_template_filepath)
-        #pprint.pprint(FS.PAGE_SPECS)
         print("-"*100)
         pprint.pprint(self.proj.framework.specs)
         print("-"*100)